requirements.txt
fastapi[all]
uvicorn[standard]
ultralytics
python-multipart
opencv-python-headless
pillow
numpy